requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.7.2
billiard==4.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-compat==1.0.15
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
greenlet==2.0.2
idna==3.4
kombu==5.3.1
MarkupSafe==2.1.3
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
